[Feat/#75] 출석 관련 API 수정 사항 반영 및 사유 제출 출석 API 연동

+++ data/src/main/java/com/umc/data/repository/schedule/ScheduleRepositoryImpl.kt
@@ -32,8 +32,8 @@ class ScheduleRepositoryImpl @Inject constructor(
     }
 
   
-    override suspend fun getScheduleDetail(scheduleId: Int): ApiState<UserCheckAvailable> {
-        return scheduleRemoteDataSource.getScheduleDetail(scheduleId.toLong()).map { response ->
+    override suspend fun getScheduleDetail(scheduleId: Long): ApiState<UserCheckAvailable> {
+        return scheduleRemoteDataSource.getScheduleDetail(scheduleId).map { response ->
             response.toModel()
         }
     }
