+++ data/src/main/java/com/umc/data/response/attendance/AttendanceAvailableResponse.kt
@@ -6,13 +6,13 @@ import com.umc.domain.model.enums.CategoryType
 import com.umc.domain.model.enums.CheckAvailableStatus
 
 data class AttendanceAvailableResponse(
-    @SerializedName("scheduleId") val scheduleId: Int,
+    @SerializedName("scheduleId") val scheduleId: Long,
     @SerializedName("scheduleName") val scheduleName: String,
     @SerializedName("tags") val tags: List<CategoryType>?,
     @SerializedName("startTime") val startTime: String,
     @SerializedName("endTime") val endTime: String,
-    @SerializedName("sheetId") val sheetId: Int,
-    @SerializedName("recordId") val recordId: Int,
+    @SerializedName("sheetId") val sheetId: Long,
+    @SerializedName("recordId") val recordId: Long,
     @SerializedName("status") val status: String,
     @SerializedName("statusDisplay") val statusDisplay: String?
 ) {
