+++ data/src/main/java/com/umc/data/response/schedule/ScheduleDetailResponse.kt
@@ -26,7 +26,7 @@ data class ScheduleDetailResponse(
         fun ScheduleDetailResponse.toModel(): UserCheckAvailable {
             return UserCheckAvailable(
                 /**TODO: 후에 Long 마이그레이션 시 해당 부분 교체 요망**/
-                id = scheduleId.toInt(),
+                id = scheduleId,
                 title = name,
                 tags = tags,
                 sheetId = 0,
