+++ domain/src/main/java/com/umc/domain/usecase/attendance/PostAttendanceApprovalUseCase.kt
@@ -4,5 +4,5 @@ import com.umc.domain.repository.attendance.AttendanceRepository
 import javax.inject.Inject
 
 class PostAttendanceApprovalUseCase @Inject constructor(private val repository: AttendanceRepository) {
-    suspend operator fun invoke(recordId: Int) = repository.approveAttendance(recordId)
+    suspend operator fun invoke(recordId: Long) = repository.approveAttendance(recordId)
 }
